spark: allows notebooks to be saved #13
There is a problem when you set a new conf folder for pyspark beyond the one by default
(this is one located in $SPARK_HOME). The problem is that you can use the loaded  jars with
the build notebooks, but you can't acces the jars from the oskar-spark notebooks. Furthermore,
when you apply "maven install", the build notebooks get erased. So I had to edit oskar-app/pom
in order to replace those notebooks at  "oskar-spark" before we erase everything in the
build folder.