spark: comment notebooks substitution code #13
There is some issue trying to connect "oskar-spark" notebooks with BUILD configuration,
and if you want to work with the BUILD notebooks the changes will be deleted whenever you
mvn install because it restores BUILD folder with his notebooks. So I made some POM CODE
in case you wantt to work there. In this situation, BUILD notebooks will be copied into
oskar-spark before it gets restored.